backend/requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.1.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Machine Learning & AI
scikit-learn==1.3.2
tensorflow==2.13.0
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2
scipy==1.11.4
xgboost==2.0.3
lightgbm==4.1.0

# Geospatial & Location Services
geopy==2.4.1
geojson==3.1.0
shapely==2.0.2
gdal==3.7.3
folium==0.15.1

# Real-time & WebSocket Support
channels==4.0.0
channels-redis==4.1.0
uvicorn[standard]==0.24.0
daphne==4.0.0

# API Documentation & Validation
drf-spectacular==0.26.5
django-filter==23.5
marshmallow==3.20.1
django-rest-swagger==2.2.0

# Storage & Cloud Services
django-storages==1.14.2
boto3==1.34.0
cloudinary==1.36.0

# Firebase Integration
firebase-admin==6.2.0
pyjwt==2.8.0
cryptography==41.0.7

# HTTP & API Clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Error Tracking & Monitoring
sentry-sdk[django]==1.39.2
django-health-check==3.17.0
django-silk==5.0.4

# Data Processing & Utilities
openpyxl==3.1.2
csv-reader==1.2.0
python-dateutil==2.8.2
pytz==2023.3
slugify==0.0.1
django-model-utils==4.3.1

# Email & Notifications
django-anymail[mailgun]==10.2
twilio==8.11.1
push-notifications==3.0.2

# Development & Testing
django-extensions==3.2.3
factory-boy==3.3.0
faker==20.1.0
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
mypy==1.7.1
django-stubs==4.2.7

# Security & Rate Limiting
django-ratelimit==4.1.0
django-axes==6.1.1
django-guardian==2.4.0

# Background Jobs & Scheduling
django-crontab==0.7.1
apscheduler==3.10.4
